pyproject.toml
[project]
name = "golgotelu"
version = "0.1.0"
description = ""
authors = [
    {name = "Laurian1906",email = "laur.laury@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai (>=1.19.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
